Move tests/test_io.py into  tests/structure/test_*.py
Matches tests with src layout
